IIPS-113 add "commonweb/WEB-INF/lib" directory to the classpath of all dependent modules
The plugin doesn't look at buildcallbacks.xml file, so it can try to compile commonweb sources together with dependent extension even if Ant build doesn't do that. If so JARs from the commonweb extension won't be copied to the dependent extension by Ant, so IDE compilation will be failed too.

Signed-off-by: Evgenii Kudelevskii ekudel@gmail.com